backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.1.0
psutil==5.9.6  # For system metrics in health checks
# boto3==1.34.0  # Optional for AWS S3 support
# redis==5.0.1   # Optional for Redis support
# celery==5.3.4  # Optional for background tasks
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
